refactor(assets): assets index 추가
